% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-I.R
\name{CalEpletMHCI}
\alias{CalEpletMHCI}
\title{calculate HLA Class I eplet mismatch using Matchmaker algorithm}
\usage{
CalEpletMHCI(dat_in)
}
\arguments{
\item{dat_in}{dataframe with subject info(first 3 columns) and MHC I allele info
each unique participant id has 2 rows associated with it, 1 for recipient, 1 for donor}
}
\value{
a list of data tables
count table: original input data appended with count mis-matched eplet of each pair
detail table: detailed mis-match eplet of each subject id, plus count and percentage of mis-match across all pairs
}
\description{
calculate HLA Class I eplet mismatch using Matchmaker algorithm
}
\examples{
\dontrun{
re <- CalEpletMHCI(dat_in = "YourDataFile")
re <- CalEpletMHCI(system.file("extdata", "MHC_I_test.csv", package = "hlaR"))
}

}
