% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHCI-test.R
\name{CalEpletMHCI_best}
\alias{CalEpletMHCI_best}
\title{many-to-many mis-match calculation, find out best match based on mm-count}
\usage{
CalEpletMHCI_best(dat_in, id_don, id_rcpt)
}
\arguments{
\item{dat_in}{data table with donor/recipient MHC-I info, each row has unique id}

\item{id_don}{donor ids}

\item{id_rcpt}{recipient ids}
}
\value{
list of data tables.
count: original input data appended with count mis-matched eplet of each pair
detail: detailed mis-match eplet of each subject id, plus count and percentage of mis-match across all pairs
}
\description{
many-to-many mis-match calculation, find out best match based on mm-count
}
\examples{
\dontrun{
dat_in <- read.csv("~/projects/DEV/hlaR_dev/MHC_I_n2n.csv", sep = ",", header = TRUE)
re <- CalEpletMHCI_best(dat_in = dat_in,
id_don = c(8), id_rcpt = c(6,1))
}

}
