% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-I_o.R, R/eplet-MHC-I_s.R
\name{CalEpletMHCI_s}
\alias{CalEpletMHCI_s}
\alias{CalEpletMHCI_o}
\title{calculate HLA Class I eplet mismatch using MatchMaker reference table and algorithm, summarized report}
\usage{
CalEpletMHCI_o(dat_in, ver = 3)

CalEpletMHCI_s(dat_in, ver = 3)
}
\arguments{
\item{dat_in}{dataframe with subject info(first 3 columns) and MHC I allele info
each unique participant id has 2 rows associated with it, 1 for recipient, 1 for donor}

\item{ver}{version number of eplet mis-match table from epitopes.net to use}
}
\value{
a list of data tables
count table: original input data appended with count mis-matched eplet of each pair
detail table: detailed mis-match eplet of each subject id, plus count and percentage of mis-match across all pairs

data table with detailed single molecule level mis-match eplet info of each subject
}
\description{
calculate HLA Class I eplet mismatch using MatchMaker reference table and algorithm, summarized report

calculate HLA Class I eplet mismatch using MatchMaker reference table and algorithm, single molecule level
}
\examples{
\dontrun{
re <- CalEpletMHCI_o(dat_in = dat, ver = 3)
}

\dontrun{
re2 <- CalEpletMHCI_s(dat, ver = 2)
re3 <- CalEpletMHCI_(dat, ver = 3)
}

}
