% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-II_old.R
\name{CalEpletMHCII_o}
\alias{CalEpletMHCII_o}
\title{calculate HLA class-II eplet mismatch using Matchmaker reference table and algorithm, summary report}
\usage{
CalEpletMHCII_o(dat_in, ver = 3)
}
\arguments{
\item{dat_in}{dataframe with subject info (first 3 columns) and MHC II allele info}

\item{ver}{version number of eplet mis-match table from epitopes.net to use}
}
\value{
a list of data tables
count table: original input data appended with count mis-matched eplet of each pair
detail table: detailed mis-match eplet of each subject id, plus count and percentage of mis-match across all pairs
}
\description{
calculate HLA class-II eplet mismatch using Matchmaker reference table and algorithm, summary report
}
\examples{
\dontrun{
re <- CalEpletMHCII_o(dat_in = dat, ver = 3)
}

}
