% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-II_old.R
\name{CalEpletMHCII_old}
\alias{CalEpletMHCII_old}
\title{calculate HLA class II eplet mismatch using Matchmaker algorithm}
\usage{
CalEpletMHCII_old(dat_in)
}
\arguments{
\item{dat_in}{dataframe with subject info (first 3 columns) and MHC II allele info}
}
\value{
a list of data tables
count table: original input data appended with count mis-matched eplet of each pair
detail table: detailed mis-match eplet of each subject id, plus count and percentage of mis-match across all pairs
}
\description{
calculate HLA class II eplet mismatch using Matchmaker algorithm
}
\examples{
\dontrun{
re <- CalEpletMHCII_old(dat_in = "YourDataFile")
re <- CalEpletMHCII_old(system.file("extdata", "MHC_II_test.csv", package = "hlaR"))
}

}
