% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-II_s.R
\name{CalEpletMHCII_s}
\alias{CalEpletMHCII_s}
\title{calculate HLA class-II eplet mismatch using Matchmaker reference table and algorithm, single molecule level}
\usage{
CalEpletMHCII_s(dat_in, ver = 3)
}
\arguments{
\item{dat_in}{dataframe with subject info (first 3 columns) and MHC II allele info}

\item{ver}{version number of eplet mis-match table from epitopes.net to use}
}
\value{
data table with detailed single molecule level mis-match eplet info of each subject
}
\description{
calculate HLA class-II eplet mismatch using Matchmaker reference table and algorithm, single molecule level
}
\examples{
\dontrun{
re2 <- CalEpletMHCII_s(dat, ver = 2)
re3 <- CalEpletMHCII_s(dat, ver = 3)
}

}
