% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla-stats.R
\name{CalMismFreq}
\alias{CalMismFreq}
\title{calculate frequency of donor mis-match alleles to recipient's}
\usage{
CalMismFreq(dat_in, nms_don = c(), nms_rcpt = c())
}
\arguments{
\item{dat_in}{dataframe of clean HLA}

\item{nms_don}{column names of donor's alleles, must be length of 2}

\item{nms_rcpt}{column names of recipient's alleles, must be length of 2}
}
\value{
dataframe of donor's mis-match alleles with frequency > 1
}
\description{
calculate frequency of donor mis-match alleles to recipient's
}
\examples{
\dontrun{
result <- CalMismFreq(dat_in = dat, nms_don = don, nms_rcpt = rcpt)
}
}
