% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-clean.R
\name{CleanAllele}
\alias{CleanAllele}
\title{clean messy HLA typing for downstream analysis}
\usage{
CleanAllele(var_1, var_2, locus)
}
\arguments{
\item{var_1}{gene on locus 1}

\item{var_2}{gene on locus 2}

\item{locus}{locus of which we want to work}
}
\value{
list of clean hla of each locus
}
\description{
clean messy HLA typing for downstream analysis
}
\examples{
\dontrun{
clean1 <- CleanAllele(clean$RECIPIENT_A1, clean$RECIPIENT_A2, locus = "a")
clean2 <- CleanAllele(clean$DONOR_DRB11, clean$DONOR_DRB11, locus = "drb")
}

}
