% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla-stats.R
\name{CountFreq}
\alias{CountFreq}
\title{Title
frequency of allele(s)}
\usage{
CountFreq(dat_in, names_in, top_n = 5)
}
\arguments{
\item{dat_in}{dataframe of clean HLA}

\item{names_in}{allele name(s) of which we want to count frequency}

\item{top_n}{names of top N most frequent alleles, default is 5}
}
\value{
dataframe of top_n most frequent alleles
}
\description{
Title
frequency of allele(s)
}
\examples{
\dontrun{
result <- CountFreq(dat_in = dat), names_in =  names, top_n = 2)
}
}
