% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla-count-mism.R
\name{CountMism}
\alias{CountMism}
\title{function to count number of mis-match HLAs}
\usage{
CountMism(dat_in, names_in)
}
\arguments{
\item{dat_in}{a dataframe with HLAs mis-match flags}

\item{names_in}{column names of which we want count mis-matches}
}
\value{
dat_in with count of mismatches appended on the last column
}
\description{
function to count number of mis-match HLAs
}
\examples{
\dontrun{
classI <- CountMism(hla, c("mism.a1", "mism.a2", "mism.b1", "mism.b2"))
classII <- CountMism(hla, c("mism.dqa12", "mism.dqb11", "mism.dqb12"  ))
}
}
