% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-eval-mism.R
\name{EvalAlleleMism}
\alias{EvalAlleleMism}
\title{Evaluate mis-match alleles.}
\usage{
EvalAlleleMism(don_1, don_2, recip_1, recip_2)
}
\arguments{
\item{don_1}{Donor's alpha1 domain.}

\item{don_2}{Donor's alpha2 or beta1 domain.}

\item{recip_1}{Recipient's alpha1 domain.}

\item{recip_2}{Recipient's alpha2 or beta1 domain.}
}
\value{
A data frame with cleaned donor and recipients, and donor to recipient mis-match flags.
}
\description{
Evaluate mis-match alleles.
}
\examples{
\dontrun{
a <- EvalAlleleMism(dat$DONOR_A1, dat$DONOR_A2, dat$RECIPIENT_A1, dat$RECIPIENT_A2)
}
}
