% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla-eval-mism.R
\name{EvalMism}
\alias{EvalMism}
\title{evaluate mis-match HLAs}
\usage{
EvalMism(dat_in, don_1, don_2, recip_1, recip_2, locus)
}
\arguments{
\item{dat_in}{a dataframe with HLAs}

\item{don_1}{donor's alpha1 domain}

\item{don_2}{donor's alpha2 or beta1 domain}

\item{recip_1}{recipient's alpha1 domain}

\item{recip_2}{recipient's alpha2 or beta1 domain}

\item{locus}{locus of domains}
}
\value{
list of mis-match vectors
mism_1 is mis-match flag of alpha1
mism_2 is mis-match flag of alpha2 or beta
}
\description{
evaluate mis-match HLAs
}
\examples{
\dontrun{
a <- EvalMism(hla, hla$donor.a1, hla$donor.a2, hla$recipient.a1, hla$recipient.a2, "a")
}
}
