% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-for-haplotype.R
\name{FuncForCompHaplo}
\alias{FuncForCompHaplo}
\alias{FuncForCompHaplo99}
\title{extracts combinations from halpo frequency table, it's called within CompHaploTbl()}
\usage{
FuncForCompHaplo99(tbl_raw, tbl_in)
}
\arguments{
\item{tbl_raw}{raw haplotype frequency table downloaded from NMDP website}

\item{tbl_in}{data frame with alleles info}
}
\description{
extracts combinations from halpo frequency table, it's called within CompHaploTbl()
}
