% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-haplotype.R
\name{ht}
\alias{ht}
\title{pull out haplotypes based on NMDP frequency table ...}
\usage{
ht(dat_in)
}
\arguments{
\item{dat_in}{dataframe of alleles}
}
\description{
pull out haplotypes based on NMDP frequency table ...
}
\examples{
\dontrun{
result <- ht(dat_in = dat)
}
}
