% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-topn.R
\name{CalAlleleTopN}
\alias{CalAlleleTopN}
\title{calculate topN most frequent allele(s)}
\usage{
CalAlleleTopN(dat_in, nms_don = c(), nms_rcpt = c(), top_n = 5)
}
\arguments{
\item{dat_in}{dataframe of clean HLA}

\item{nms_don}{donor's allele name(s) frequency count}

\item{nms_rcpt}{recipient's allele name(s) of frequency count}

\item{top_n}{names of top N most frequent alleles, default is 5}
}
\value{
dataframe of top_n most frequent alleles
}
\description{
calculate topN most frequent allele(s)
}
\examples{
\dontrun{
result <- CalAlleleTopN(dat_in = dat, nms_don = don, nms_rcpt = rcpt, top_n = 2)
}
}
