% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-I.R
\name{CalEpletMHCI}
\alias{CalEpletMHCI}
\title{Calculate HLA Class I eplet mismatch. Eplet Reference table and mistmatch calculation logic are based on MatchMaker.}
\usage{
CalEpletMHCI(dat_in, ver = 3)
}
\arguments{
\item{dat_in}{A dataframe with subject info(first 2 columns) and MHC I allele info.
Each unique participant id has 2 rows associated with it, 1 for recipient, 1 for donor.}

\item{ver}{Version number of MatchMaker based eplet reference to use.}
}
\value{
A list of data tables.
single_detail: single moelcule level eplet mismatch table, including mismatch eplet name and count of each allele
overall_count: original input data appended with count of unique mis-matched eplet
overall_detail: percentage of mismatch across all subject of each eplet
}
\description{
Calculate HLA Class I eplet mismatch. Eplet Reference table and mistmatch calculation logic are based on MatchMaker.
}
\examples{
\dontrun{
re <- CalEpletMHCI(dat_in = dat, ver = 3)
}

}
