% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-I.R
\name{CalEpletMHCI}
\alias{CalEpletMHCI}
\title{calculate HLA Class I eplet mismatch using Matchmaker algorithm}
\usage{
CalEpletMHCI(dat_in, ver)
}
\arguments{
\item{dat_in}{dataframe with subject info(first 3 columns) and MHC I allele info
each unique participant id has 2 rows associated with it, 1 for recipient, 1 for donor}

\item{ver}{version of eplet mis-match table from epitopes.net}
}
\value{
data table with detailed single molecule level mis-match eplet info of each subject
}
\description{
calculate HLA Class I eplet mismatch using Matchmaker algorithm
}
\examples{
\dontrun{
re <- CalEpletMHCI(dat_in = "YourDataFile", ver = 2)
re <- CalEpletMHCI(system.file("extdata", "MHC_I_test.csv", package = "hlaR"), ver = 2)
}

}
