% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-I.R
\name{CalEpletMHCI}
\alias{CalEpletMHCI}
\title{calculate HLA Class I eplet mismatch using MatchMaker reference table and algorithm, summary report}
\usage{
CalEpletMHCI(dat_in, ver = 3)
}
\arguments{
\item{dat_in}{dataframe with subject info(first 2 columns) and MHC I allele info
each unique participant id has 2 rows associated with it, 1 for recipient, 1 for donor}

\item{ver}{version number of eplet mis-match table from epitopes.net to use}
}
\value{
a list of data tables
single_detail: single moelcule level eplet mismatch table, including mismatch eplet name and count of each allele
overall_count: original input data appended with count of unique mis-matched eplet
overall_detail: percentage of mismatch across all subject of each eplet
}
\description{
calculate HLA Class I eplet mismatch using MatchMaker reference table and algorithm, summary report
}
\examples{
\dontrun{
re <- CalEpletMHCI(dat_in = dat, ver = 3)
}

}
