% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-I.R
\name{CalEpletMHCI}
\alias{CalEpletMHCI}
\title{Calculate class I MHC eplet mismatch.}
\usage{
CalEpletMHCI(dat_in, ver = 3)
}
\arguments{
\item{dat_in}{A dataframe with subject info (pair_id and subject_type) and both copies of each class I locus (HLA A, B, and C).
Each unique pair_id has 2 rows associated with it, 1 for recipient, 1 for donor.}

\item{ver}{Version number of MatchMaker based eplet reference table to use.}
}
\value{
A list of data tables.
single_detail: single molecule class I MHC eplet mismatch table, including mismatched eplet names and the count of eplets mismatched at each allele.
overall_count: original input data appended with total count of mismatched eplets.
overall_detail: percentage of mismatch across all subjects for each eplet.
}
\description{
Use high resolution HLA class I data to calculate class I eplet mismatch for a population of donors and recipients. Mismatch is calculated using logic from HLAMatchMaker, developed by Rene Dusquesnoy. Current reference tables supported are HLAMatchMaker v2 and v3.
}
\examples{
\dontrun{
re <- CalEpletMHCI(dat_in = dat, ver = 3)
}

}
