% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-II.R, R/eplet-MHC-II_old.R
\name{CalEpletMHCII}
\alias{CalEpletMHCII}
\title{calculate HLA class II eplet mismatch using Matchmaker algorithm}
\usage{
CalEpletMHCII(dat_in)

CalEpletMHCII(dat_in)
}
\arguments{
\item{dat_in}{dataframe with subject info (first 3 columns) and MHC II allele info}
}
\value{
data table with detailed single molecular level mis-match eplet info of each subject

a list of data tables
count table: original input data appended with count mis-matched eplet of each pair
detail table: detailed mis-match eplet of each subject id, plus count and percentage of mis-match across all pairs
}
\description{
calculate HLA class II eplet mismatch using Matchmaker algorithm

calculate HLA class II eplet mismatch using Matchmaker algorithm
}
\examples{
\dontrun{
re <- CalEpletMHCII(dat_in = "YourDataFile")
re <- CalEpletMHCII(system.file("extdata", "MHC_II_test.csv", package = "hlaR"))
}

\dontrun{
re <- CalEpletMHCII(dat_in = "YourDataFile")
re <- CalEpletMHCII(system.file("extdata", "MHC_II_test.csv", package = "hlaR"))
}

}
