% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-II.R
\name{CalEpletMHCII}
\alias{CalEpletMHCII}
\title{Calculate class II HLA eplet mismatch. Eplet Reference table and mismatch calculation logic are based on MatchMaker. Note: interlocus info only available in table v3 reference tables.}
\usage{
CalEpletMHCII(dat_in, ver = 3)
}
\arguments{
\item{dat_in}{A dataframe with subject info (first 2 columns) and class II MHC allele info.}

\item{ver}{Version number of MatchMaker based class II MHC eplet reference table to use. Mismatch calculation logic are based on MatchMaker too.}
}
\value{
A list of data tables.
single_detail: single molecule level eplet mismatch table, including mismatch eplet name and count of each allele.
overall_count: original input data appended with count of unique mismatched eplet.
}
\description{
Calculate class II HLA eplet mismatch. Eplet Reference table and mismatch calculation logic are based on MatchMaker. Note: interlocus info only available in table v3 reference tables.
}
\examples{
\dontrun{
re2 <- CalEpletMHCII(dat, ver = 2)
re3 <- CalEpletMHCII(dat, ver = 3)
}

}
