% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-II.R
\name{CalEpletMHCII}
\alias{CalEpletMHCII}
\title{calculate HLA class-II eplet mismatch using Matchmaker reference table and algorithm, single molecule level}
\usage{
CalEpletMHCII(dat_in, ver = 3)
}
\arguments{
\item{dat_in}{a dataframe with subject info (first s columns) and MHC II allele info}

\item{ver}{version number of eplet mis-match table from epitopes.net to use}
}
\value{
a list of data tables
single_detail: single molecule level eplet mismatch table, including mismatch eplet name and count of each allele
overall_count: original input data appended with count of unique mis-matched eplet
}
\description{
calculate HLA class-II eplet mismatch using Matchmaker reference table and algorithm, single molecule level
}
\examples{
\dontrun{
re2 <- CalEpletMHCII(dat, ver = 2)
re3 <- CalEpletMHCII(dat, ver = 3)
}

}
