% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-II.R
\name{CalEpletMHCII}
\alias{CalEpletMHCII}
\title{calculate HLA class II eplet mismatch using Matchmaker algorithm}
\usage{
CalEpletMHCII(dat_in)
}
\arguments{
\item{dat_in}{dataframe with subject info (first 3 columns) and MHC II allele info}
}
\value{
data table with detailed single molecular level mis-match eplet info of each subject
}
\description{
calculate HLA class II eplet mismatch using Matchmaker algorithm
}
\examples{
\dontrun{
re <- CalEpletMHCII(dat_in = "YourDataFile")
re <- CalEpletMHCII(system.file("extdata", "MHC_II_test.csv", package = "hlaR"))
}

}
