% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-II.R
\name{CalEpletMHCII}
\alias{CalEpletMHCII}
\title{calculate HLA class II eplet mismatch using Matchmaker algorithm}
\usage{
CalEpletMHCII(dat_in)
}
\arguments{
\item{dat_in}{input data set with or without complete allele info
it has 27 columns (first 3 columns are record id, recipient id, donor id; the rest of columns are MHC II alleles for recipient and donor)}
}
\value{
list of data tables.
count: original input data appended with count mis-matched eplet of each pair
detail: detailed mis-match eplet of each subject id, plus count and percentage of mis-match across all pairs
}
\description{
calculate HLA class II eplet mismatch using Matchmaker algorithm
}
\examples{
\dontrun{
re <- CalEpletMHCII(dat_in = "YourDataFile")
re <- CalEpletMHCII(system.file("extdata", "MHC_II_test.csv", package = "hlaR"))
}

}
