% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eplet-MHC-II.R
\name{CalEpletMHCII}
\alias{CalEpletMHCII}
\title{Calculate HLA class-II eplet mismatch. Eplet Reference table and mistmatch calculation logic are based on MatchMaker. Note: interlocus info only available in table v3 reference tables.}
\usage{
CalEpletMHCII(dat_in, ver = 3)
}
\arguments{
\item{dat_in}{A dataframe with subject info (first 2 columns) and MHC II allele info.}

\item{ver}{Version number of MatchMaker based eplet reference to use.}
}
\value{
A list of data tables.
single_detail: single molecule level eplet mismatch table, including mismatch eplet name and count of each allele
overall_count: original input data appended with count of unique mis-matched eplet
}
\description{
Calculate HLA class-II eplet mismatch. Eplet Reference table and mistmatch calculation logic are based on MatchMaker. Note: interlocus info only available in table v3 reference tables.
}
\examples{
\dontrun{
re2 <- CalEpletMHCII(dat, ver = 2)
re3 <- CalEpletMHCII(dat, ver = 3)
}

}
