% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla-stats.R
\name{CalMismFreq}
\alias{CalMismFreq}
\title{calculate mis-match frequency of donor's alleles to recipient's}
\usage{
CalMismFreq(dat_in, names_don, names_rcpt)
}
\arguments{
\item{dat_in}{dataframe of clean HLA}

\item{names_don}{column names of donor's alleles, must be length of 2}

\item{names_rcpt}{column names of recipient's alleles, must be length of 2}
}
\value{
dataframe of donor's mis-match alleles with frequent > 1
}
\description{
calculate mis-match frequency of donor's alleles to recipient's
}
\examples{
\dontrun{
result <- CalMismFreq(dat_in = dat, names_don = don, names_rcpt = rcpt)
}



}
