% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-clean.R
\name{CleanAllele}
\alias{CleanAllele}
\title{Clean messy HLA typing data}
\usage{
CleanAllele(var_1, var_2)
}
\arguments{
\item{var_1}{HLA on allele 1.}

\item{var_2}{HLA on allele 2.}
}
\value{
A data frame of cleaned HLA data for each loci.
}
\description{
This function takes raw “messy” HLA typing data as input. It removes inconsistent formatting and unnecessary symbols. If one of two alleles at a loci is NA, the locus is assumed to be homozygous.
}
\examples{
\dontrun{
clean1 <- CleanAllele(clean$RECIPIENT_A1, clean$RECIPIENT_A2)
clean2 <- CleanAllele(clean$DONOR_DRB11, clean$DONOR_DRB11)
}

}
