% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-clean.R
\name{CleanAllele}
\alias{CleanAllele}
\title{Clean messy HLA typing for downstream analysis. For NA alleles, if both loci are not NA or both are NA, then keep them as they are; if one of loci is NA, then assign the non-NA one to it.}
\usage{
CleanAllele(var_1, var_2)
}
\arguments{
\item{var_1}{Hla on locus 1.}

\item{var_2}{Hla on locus 2.}
}
\value{
A data frame with cleaned hla of each locus.
}
\description{
Clean messy HLA typing for downstream analysis. For NA alleles, if both loci are not NA or both are NA, then keep them as they are; if one of loci is NA, then assign the non-NA one to it.
}
\examples{
\dontrun{
clean1 <- CleanAllele(clean$RECIPIENT_A1, clean$RECIPIENT_A2)
clean2 <- CleanAllele(clean$DONOR_DRB11, clean$DONOR_DRB11)
}

}
