% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-clean.R
\name{CleanAllele}
\alias{CleanAllele}
\title{clean messy HLA typing for downstream analysis}
\usage{
CleanAllele(var_1, var_2, locus)
}
\arguments{
\item{var_1}{alpha1 domain we want to clean}

\item{var_2}{alpha2 or beta1 domain we want to clean}

\item{locus}{locus of which we want to work
var_1 is hla of alpha1(or alpha for class II) allele
var_2 is hla of alpha2(or beta for class II) allele
locus is indicator of which locus we are working on}
}
\value{
list of data cleaned hla of the locus
returns alpha1 and alpha2 for HLA class I
returns alpha and beta for HLA class II
}
\description{
clean messy HLA typing for downstream analysis
}
\examples{
\dontrun{
re <- CleanAllele(dat$don.a1.ori, dat$don.a2.ori, locus = "a")
re <- CleanAllele(dat$don.drb.ori, dat$don.drb.ori, locus = "drb")
}

}
