% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-clean.R
\name{CleanAllele}
\alias{CleanAllele}
\title{clean messy HLA typing for downstream analysis}
\usage{
CleanAllele(var_1, var_2)
}
\arguments{
\item{var_1}{hla on locus 1}

\item{var_2}{hla on locus 2}
}
\value{
a data frame with cleaned hla of each locus

a data frame with cleaned hla of each locus
}
\description{
clean messy HLA typing for downstream analysis

clean messy HLA typing for downstream analysis
}
\examples{
\dontrun{
clean1 <- CleanAllele(clean$RECIPIENT_A1, clean$RECIPIENT_A2)
clean2 <- CleanAllele(clean$DONOR_DRB11, clean$DONOR_DRB11)
}

\dontrun{
clean1 <- CleanAllele(clean$RECIPIENT_A1, clean$RECIPIENT_A2)
clean2 <- CleanAllele(clean$DONOR_DRB11, clean$DONOR_DRB11)
}

}
