% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-haplotype.R
\name{CompHaploTbl}
\alias{CompHaploTbl}
\title{compare alleles with NMDP frequency table to get top combination of haplotypes based on user cutoffs}
\usage{
CompHaploTbl(dat_in, cut_p, cut_r)
}
\arguments{
\item{dat_in}{dataframe with recipient/donor alleles info}

\item{cut_p}{cutoff frequence value of the population, it passed to FuncForCompHaplo()}

\item{cut_r}{cutoff rank value of each count group, it passed to FuncForCompHaplo()}
}
\value{
a list of dataframe of most matched allele combination of each subject
}
\description{
compare alleles with NMDP frequency table to get top combination of haplotypes based on user cutoffs
}
\examples{
\dontrun{
result <- CompHaploTbl(dat_in = dat, cut_p= 0.0001, cut_r = 10)
}
}
