% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-haplotype.R
\name{CompHaploTbl}
\alias{CompHaploTbl}
\title{compare alleles with NMDP frequency table to get top combination of haplotypes based on user cutoffs}
\usage{
CompHaploTbl(dat_in)
}
\arguments{
\item{dat_in}{dataframe with recipient/donor alleles info}
}
\value{
a list of dataframe of most matched allele combination of each subject
}
\description{
compare alleles with NMDP frequency table to get top combination of haplotypes based on user cutoffs
}
\examples{
\dontrun{
result <- CompHaploTbl(dat_in = dat, cut_p= 0.0001, cut_n = 10)
}
}
