% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-count-mism.R
\name{CountAlleleMism}
\alias{CountAlleleMism}
\title{count number of mis-match alleles}
\usage{
CountAlleleMism(dat_in, names_in)
}
\arguments{
\item{dat_in}{a dataframe with HLAs mis-match flags}

\item{names_in}{column names of which we want count mis-matches}
}
\value{
dat_in with count of mismatches appended on the last column
}
\description{
count number of mis-match alleles
}
\examples{
\dontrun{
classI <- CountAlleleMism(hla, c("mism.a1", "mism.a2", "mism.b1", "mism.b2"))
classII <- CountAlleleMism(hla, c("mism.dqa12", "mism.dqb11", "mism.dqb12"  ))
}
}
