% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-count-mism.R
\name{CountAlleleMism}
\alias{CountAlleleMism}
\title{Count number of mis-match alleles. Count is NA if input alleles are NA.}
\usage{
CountAlleleMism(dat_in, names_in)
}
\arguments{
\item{dat_in}{A dataframe with allele's mis-match flag.}

\item{names_in}{A column names to count mis-matche.}
}
\value{
dat_in with count of mismatches appended on the last column
}
\description{
Count number of mis-match alleles. Count is NA if input alleles are NA.
}
\examples{
\dontrun{
classI <- CountAlleleMism(hla_mm_cnt, c("mism.a1", "mism.a2", "mism.b1", "mism.b2"))
classII <- CountAlleleMism(hla_mm_cnt, c("mism.dqa12", "mism.dqb11", "mism.dqb12"))
}
}
