% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-count-mism.R
\name{CountAlleleMism}
\alias{CountAlleleMism}
\title{Count HLA mismatch at the allele level}
\usage{
CountAlleleMism(dat_in, names_in)
}
\arguments{
\item{dat_in}{A dataframe with donor and recipient mismatched alleles as output from EvalAlleleMism function.}

\item{names_in}{A column containing the names of HLA loci for which to count mismatch.}
}
\value{
The input data, 'dat_in' is returned with the count of mismatches appended as the last column.
}
\description{
Donor and recipient HLA typing data is compared to determine allele level mismatch. The output of EvalAlleleMism is used as input for this function. Allele level mismatch can be calculated for both high and low resolution data. The generated count will return 'NA' if the input alleles are 'NA.'
}
\examples{
\dontrun{
classI <- CountAlleleMism(hla_mm_cnt, c("mism.a1", "mism.a2", "mism.b1", "mism.b2"))
classII <- CountAlleleMism(hla_mm_cnt, c("mism.dqa12", "mism.dqb11", "mism.dqb12"))
}
}
