% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-eval-mism.R
\name{EvalAlleleMism}
\alias{EvalAlleleMism}
\title{evaluate mis-match alleles}
\usage{
EvalAlleleMism(dat_in, don_1, don_2, recip_1, recip_2)
}
\arguments{
\item{dat_in}{dataframe with allele info}

\item{don_1}{donor's alpha1 domain}

\item{don_2}{donor's alpha2 or beta1 domain}

\item{recip_1}{recipient's alpha1 domain}

\item{recip_2}{recipient's alpha2 or beta1 domain}
}
\value{
a list of mis-match vectors
mism_1 is mis-match flag of locus 1
mism_2 is mis-match flag of locus 2
}
\description{
evaluate mis-match alleles
}
\examples{
\dontrun{
a <- EvalAlleleMism(dat, dat$donor.a1, dat$donor.a2, dat$recipient.a1, dat$recipient.a2)
}
}
