% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-eval-mism.R
\name{EvalAlleleMism}
\alias{EvalAlleleMism}
\title{evaluate mis-match alleles}
\usage{
EvalAlleleMism(don_1, don_2, recip_1, recip_2)
}
\arguments{
\item{don_1}{donor's alpha1 domain}

\item{don_2}{donor's alpha2 or beta1 domain}

\item{recip_1}{recipient's alpha1 domain}

\item{recip_2}{recipient's alpha2 or beta1 domain}
}
\value{
a data frame with cleaned donor and recipients, and mis-match flags on low and high resolution alleles of donor to recipient
}
\description{
evaluate mis-match alleles
}
\examples{
\dontrun{
a <- EvalAlleleMism99(dat$DONOR_A1, dat$DONOR_A2, dat$RECIPIENT_A1, dat$RECIPIENT_A2)
}
}
