% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-eval-mism.R
\name{EvalAlleleMism}
\alias{EvalAlleleMism}
\title{Evaluate mismatched alleles}
\usage{
EvalAlleleMism(don_1, don_2, recip_1, recip_2, hmz_cnt = 1)
}
\arguments{
\item{don_1}{Donor's alpha1 domain.}

\item{don_2}{Donor's alpha2 or beta1 domain.}

\item{recip_1}{Recipient's alpha1 domain.}

\item{recip_2}{Recipient's alpha2 or beta1 domain.}

\item{hmz_cnt}{Use hmz_cnt to determine how mismatch at homozygous alleles should be handled. By default, a mismatch at a homozygous allele is considered a single mismatch. Set hmz_cnt = 2 to count homozygous mismatches as double.}
}
\value{
A data frame of cleaned donor and recipient HLA typing data, with mismatched donor-to-recipient alleles flagged.
}
\description{
Compare donor and recipient HLA typing data to determine mismatched alleles. Input data can be high or low resolution, mismatch is evaluated at the allele level.
}
\examples{
\dontrun{
a <- EvalAlleleMism(dat$DONOR_A1, dat$DONOR_A2, dat$RECIPIENT_A1, dat$RECIPIENT_A2, hmz_cnt = 2)
}
}
