% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-for-haplotype.R
\name{FuncForCompHaplo}
\alias{FuncForCompHaplo}
\title{Compare Haplotypes}
\usage{
FuncForCompHaplo(tbl_raw, tbl_in)
}
\arguments{
\item{tbl_raw}{A raw haplotype frequency table downloaded from NMDP website.}

\item{tbl_in}{A dataframe with alleles info.}
}
\description{
This function generates the ranked haplotypes for high resolution imputation. It is called by the ImputeHaplo() function.
}
