% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-for-haplotype.R
\name{FuncForCompHaplo}
\alias{FuncForCompHaplo}
\title{Compare Haplotypes}
\usage{
FuncForCompHaplo(tbl_raw, tbl_in)
}
\arguments{
\item{tbl_raw}{A raw haplotype frequency table provided by National Marrow Donor Program.}

\item{tbl_in}{A data frame with low or mixed resolution HLA typing.}
}
\value{
A data frame of imputed high resolution haplotypes.
}
\description{
This function generates the ranked haplotypes for high resolution imputation. It is called within the ImputeHaplo function.
}
\examples{
\dontrun{
for (i in 1:num_subj){
re[[i]] <- FuncForCompHaplo(tbl_raw = raw_hap_tbl, tbl_in = dat[i, ])
}
}
}
