% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-for-haplotype.R
\name{FuncForCompHaplo}
\alias{FuncForCompHaplo}
\title{extracts combinations from halpo frequency table, it's called within CompHaploTbl()}
\usage{
FuncForCompHaplo(tbl_in1, tbl_in2, cut_freq = 1e-04, cut_rank = 10)
}
\arguments{
\item{tbl_in1}{raw haplotype frequency table downloaded from NMDP website}

\item{tbl_in2}{tbl_in2aframe of alleles}

\item{cut_freq}{cutoff frequence value of the population}

\item{cut_rank}{cutoff rank value of each count group}
}
\description{
extracts combinations from halpo frequency table, it's called within CompHaploTbl()
}
