% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-haplotype.R
\name{ImputeHaplo}
\alias{ImputeHaplo}
\title{Impute low resolution hla to high resolution typing.}
\usage{
ImputeHaplo(dat_in)
}
\arguments{
\item{dat_in}{A data frame with recipient/donor alleles info.}
}
\value{
A data frame with the best pairs of haplotype combination.
}
\description{
Impute low resolution hla to high resolution typing.
}
\examples{
\dontrun{
result <- ImputeHaplo(dat_in = dat)
}
}
