% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele-haplotype.R
\name{ImputeHaplo}
\alias{ImputeHaplo}
\title{feed low resolution hla typing to NMDP frequency table to get best pairs of haplotype combination based on max count of unique low res in the combination}
\usage{
ImputeHaplo(dat_in)
}
\arguments{
\item{dat_in}{dataframe with recipient/donor alleles info}
}
\value{
dataframe of best pairs of haplotype combination
}
\description{
feed low resolution hla typing to NMDP frequency table to get best pairs of haplotype combination based on max count of unique low res in the combination
}
\examples{
\dontrun{
result <- ImputeHaplo(dat_in = dat)
}
}
