% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{GenerateLookup}
\alias{CalRiskScore}
\alias{FuncForCompHaplo}
\alias{na_to_empty_string}
\title{Basic functions}
\usage{
GenerateLookup(lkup_in, locus_in)

CalRiskScore(dat_in)

FuncForCompHaplo(tbl_raw, tbl_in)

na_to_empty_string(df)
}
\arguments{
\item{lkup_in}{data table}

\item{locus_in}{string
CalRiskScore() calculate DR DQ risk score, it's called in CalEpletMHCII()}

\item{dat_in}{dataframe
FuncForCompHaplo() called in ImputeHaplo()}

\item{tbl_raw}{data frame}

\item{tbl_in}{data frame
na_to_empty_string()}

\item{df}{dataframe}
}
\description{
GenerateLookup() called in CalEpletMHCII()
}
